+++ upcloud/network_test.go
@@ -187,6 +187,10 @@ func TestUnmarshalNetwork(t *testing.T) {
 				  "172.16.0.10",
 				  "172.16.1.10"
 				],
+				"dhcp_routes" : [
+				  "192.168.0.0/24",
+				  "192.168.100.100/32"
+				],
 				"family": "IPv4",
 				"gateway": "172.16.0.1"
 			  }
@@ -230,6 +234,10 @@ func TestUnmarshalNetwork(t *testing.T) {
 				DHCPDns: []string{
 					"172.16.0.10",
 					"172.16.1.10",
+				},
+				DHCPRoutes: []string{
+					"192.168.0.0/24",
+					"192.168.100.100/32",
 				},
 				Family:  IPAddressFamilyIPv4,
 				Gateway: "172.16.0.1",