+++ CHANGELOG.md
@@ -5,6 +5,8 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [6.7.0]
+
 ### Added
 - network: `dhcp_routes` field to IP network for additional DHCP classless static routes to be delivered if the DHCP is enabled
 - network: `static_routes` field to router for defining static routes
@@ -338,7 +340,8 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 First stable release
 
-[Unreleased]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v6.6.0...HEAD
+[Unreleased]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v6.7.0...HEAD
+[6.7.0]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v6.6.0...v6.7.0
 [6.6.0]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v6.5.0...v6.6.0
 [6.5.0]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v6.4.0...v6.5.0
 [6.4.0]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v6.3.2...v6.4.0